% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{check_fail}
\alias{check_fail}
\title{Determine if problem occured during downloading of an issue}
\usage{
check_fail(x)
}
\arguments{
\item{x}{dataframe return of chronam::get_issues}
}
\value{
bool
}
\description{
Determine if problem occured during downloading of an issue
}
