% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content.R
\name{download_batch_ocr_tar}
\alias{download_batch_ocr_tar}
\title{Download the batch ocr data from chronicling America}
\usage{
download_batch_ocr_tar(batch_name, odir = ".")
}
\arguments{
\item{batch_name}{character name of batch e.g wvu_els_ver01}

\item{odir}{optional character path of directory to download the files into}
}
\description{
bulk data url: http://chroniclingamerica.loc.gov/data/batches/<batch_name>
bulk ocr data url: http://chroniclingamerica.loc.gov/data/ocr/<batch_name>.tar.bz2
}
\details{
batch info url: http://chroniclingamerica.loc.gov/batches/<batch_name>.json
bulk bib data url: http://chroniclingamerica.loc.gov/data/bib/<??>
bulk word_coordinates data url: http://chroniclingamerica.loc.gov/data/word_coordinates/<lccn>

the ocr data url seems be missing batches ingested in the current month
}
