% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content.R
\name{exract_xml_files}
\alias{exract_xml_files}
\title{Extract ocr xml files from tar file}
\usage{
exract_xml_files(path_to_tar, exdir = ".")
}
\arguments{
\item{path_to_tar}{character file path to tar file for a batch}

\item{exdir}{character file path extract contents to}
}
\description{
Extract ocr xml files from tar file
}
