% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{get_batch}
\alias{get_batch}
\title{Get Batch for each issue}
\usage{
get_batch(issue_url)
}
\arguments{
\item{issue_url}{character url for issue metadata}
}
\value{
character of batch name
}
\description{
Get Batch for each issue
}
