% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content.R
\name{get_batch_from_issue}
\alias{get_batch_from_issue}
\title{Get Batch for each issue}
\usage{
get_batch_from_issue(issue_url)
}
\arguments{
\item{issue_url}{character url for issue metadata}
}
\value{
character of batch name
}
\description{
It would have been faster to lookup based on lccn and chronam::batches
Have to do it this way because each lccn can appear in multiple batches
Since this is very slow, only run this when downloading the data for
a set of the issues that is much smaller than the 2.6mill+ in the full
collection. I.e only use this when you have a sample
}
