% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batches.R
\name{get_batches}
\alias{get_batches}
\title{Get full list of batch metadata from Chronicling America Website}
\usage{
get_batches()
}
\value{
tibble of all the batches
}
\description{
Takes a couple of minutes.
}
