% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content.R
\name{get_block_content}
\alias{get_block_content}
\title{Aggregate all the strings in a text block}
\usage{
get_block_content(block)
}
\arguments{
\item{xmlnode}{for a TextBlock from ocr.xml file}
}
\value{
string containing all the words in the text block
}
\description{
Aggregate all the strings in a text block
}
