% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{get_issues}
\alias{get_issues}
\title{Get the issues associated with a newspaper}
\usage{
get_issues(lccn)
}
\arguments{
\item{lccn}{character the loc control number for the newspaper}
}
\value{
tibble of metadata for each issue of that newspaper
}
\description{
Get the issues associated with a newspaper
}
