% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{get_issues_all}
\alias{get_issues_all}
\title{Get a dataframe of all the issues and their metadata}
\usage{
get_issues_all(sample = -1, ncores = parallel::detectCores())
}
\arguments{
\item{sample}{optional int refering to sample size, by default gets all}

\item{ncores}{optional int number of cores to use for pinging chronam site}
}
\value{
dataframe of all the issues and their metadata
}
\description{
this takes about 30 minutes to run on one core
Use package data when possible
}
