% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{get_newspapers}
\alias{get_newspapers}
\title{Get full list of newspapers metadata from Chronicling America Website}
\usage{
get_newspapers()
}
\value{
tibble of all the newspapers
}
\description{
Get full list of newspapers metadata from Chronicling America Website
}
