% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{get_state}
\alias{get_state}
\title{Get the state assignment for a each issue}
\usage{
get_state(place)
}
\arguments{
\item{place}{character of locations associated with the newspapers}
}
\value{
character the state name for the last place in the list
}
\description{
Get the state assignment for a each issue
}
