% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{issues_in_range}
\alias{issues_in_range}
\title{Return metadata about newpapers issues from within range}
\usage{
issues_in_range(
  start,
  end,
  issues = chronam::issues,
  newspapers = chronam::newspapers
)
}
\arguments{
\item{start}{number for inclusive start date}

\item{end}{number for inclusive end date}

\item{issues}{optional dataframe of metadata}
}
\value{
tibble of metadata for each issue of that newspaper that has issues\
within the start and end date
}
\description{
Return metadata about newpapers issues from within range
}
