% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content.R
\name{parse_batch_ocr_tar}
\alias{parse_batch_ocr_tar}
\title{Load the text for a locally downloaded batch}
\usage{
parse_batch_ocr_tar(path_to_tar, exdir = ".")
}
\arguments{
\item{path_to_tar}{character file path to tar file for a batch}

\item{exdir}{character file path extract contents to}
}
\value{
dataframe of text_block, page, issue, date, year, month, day, lccn
}
\description{
Load the text for a locally downloaded batch
}
