% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batches.R
\name{parse_batches_part_json}
\alias{parse_batches_part_json}
\title{Get metadata from a single json page of batch results}
\usage{
parse_batches_part_json(url)
}
\arguments{
\item{url}{character url of the part to get data from}
}
\value{
list with nextcursor url (can be null) and batches dataframe
}
\description{
Get metadata from a single json page of batch results
}
