% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content.R
\name{parse_ocr_xml}
\alias{parse_ocr_xml}
\title{Parse the xml from an ocr.xml file}
\usage{
parse_ocr_xml(xml, root = "")
}
\arguments{
\item{xml}{character filepath or xml content of an ocr.xml file}
}
\value{
tibble where each row is a text block
}
\description{
Parse the xml from an ocr.xml file
}
