% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content.R
\name{parse_xml_files}
\alias{parse_xml_files}
\title{Parse directory of xml files into dataframe}
\usage{
parse_xml_files(path)
}
\arguments{
\item{path}{to a directory of xml files}
}
\value{
dataframe of text_block, page, issue, date, year, month, day, lccn
}
\description{
Parse directory of xml files into dataframe
}
