% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{get_issues_all}
\alias{get_issues_all}
\title{Get a dataframe of all the issues and their metadata}
\usage{
get_issues_all(sample = -1, ncores = detectCores(), update = FALSE)
}
\arguments{
\item{sample}{optional int refering to sample size, by default gets all}

\item{ncores}{optional int number of cores to use for pinging chronam site}

\item{update}{optional bool if FALSE just use the package data}
}
\value{
dataframe of all the issues and their metadata
}
\description{
this takes about 30 minutes to run on one core
so by default we just load the package data
if you suspect package data is out of date then
can be run with update=TRUE param
}
