% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset_info}
\alias{subset_info}
\title{Return metadata about newpapers issues}
\usage{
subset_info(issues, newspapers = chronam::newspapers)
}
\arguments{
\item{issues}{dataframe of issues to summarize information for}

\item{newspapers}{optional dataframe of newspaper metadata}
}
\value{
tibble of metadata for each issue of that newspaper that has issues\
within the start and end date
}
\description{
Return metadata about newpapers issues
}
